Changes the package name from 'process.model' to 'model'.
