+++ bundles/orbistoolbox/src/test/java/org/orbisgis/orbistoolbox/model/DescriptionTypeTest.java
@@ -17,12 +17,10 @@
  * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
  */
 
-package org.orbisgis.orbistoolbox.process.model;
+package org.orbisgis.orbistoolbox.model;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.orbisgis.orbistoolbox.process.model.DescriptionType;
-import org.orbisgis.orbistoolbox.process.model.Metadata;
 
 import java.net.URI;
 import java.net.URISyntaxException;
