+++ bundles/orbistoolbox/src/test/java/org/orbisgis/orbistoolbox/model/OutputTest.java
@@ -17,14 +17,10 @@
  * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
  */
 
-package org.orbisgis.orbistoolbox.process.model;
+package org.orbisgis.orbistoolbox.model;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.orbisgis.orbistoolbox.process.model.DataDescription;
-import org.orbisgis.orbistoolbox.process.model.Format;
-import org.orbisgis.orbistoolbox.process.model.Output;
-import org.orbisgis.orbistoolbox.process.model.RawData;
 
 import java.net.URI;
 import java.net.URISyntaxException;
